Renamed command files to represent their place in the apio command hirearchy. E.g. board.py -> apio_board.py.  This change regresses apio -h, and some of the test and this will be addressed in a follow up cl that sets a proper click command  hirerachy.
