+++ test/integration/test_examples.py
@@ -8,8 +8,8 @@ from test.conftest import ApioRunner
 import pytest
 
 # -- Entry point for apio commands.
-from apio.commands.packages import cli as apio_packages
-from apio.commands.examples import cli as apio_examples
+from apio.commands.apio_packages import cli as apio_packages
+from apio.commands.apio_examples import cli as apio_examples
 
 
 # R0801: Similar lines in 2 files
